HW-2-RAG/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "pydantic>=2.11.7",
    "uvicorn>=0.34.3",
    "vectorize-client>=0.2.1",
    "openai>=1.41.2",
    "python-dotenv>=1.0.1",
    "litellm>=1.40.10",
]
